refactor more components
