+++ components/Header.js
@@ -1,7 +1,7 @@
 import { Box, Flex, HStack, Link } from "./elements";
 import { IconTwitch, IconYoutube } from "./icons";
 import { Logo } from "./Logo";
-import { NavigationLinks } from "./NavigationLinks";
+import { SectionLinks } from "./SectionLinks";
 
 export const Header = () => {
   return (
@@ -13,10 +13,10 @@ export const Header = () => {
         justify="space-between"
         align="center"
       >
-        <h1 title="Desarrollo avaScript Fullstack - Carlos Azaustre">
+        <h1 title="Desarrollo JavaScript Fullstack - Carlos Azaustre">
           <Logo />
         </h1>
-        <NavigationLinks />
+        <SectionLinks />
         <HStack spacing="40px">
           <Link
             href="//youtube.com/c/CarlosAzaustre?sub_confirmation=1"
