+++ components/PostMetadata.js
@@ -1,8 +1,8 @@
 import { formatDate } from "@/lib/format-date";
-import { Flex, Text, Tag, Link } from "../elements";
-import { IconCalendar, IconComment, IconClock } from "../icons";
-import { OpenPR } from "../OpenPR";
-import { Disqus } from "../Disqus";
+import { Flex, Text, Tag, Link } from "./elements";
+import { IconCalendar, IconComment, IconClock } from "./icons";
+import { OpenPR } from "./OpenPR";
+import { Disqus } from "./Disqus";
 
 export const PostMetadata = ({ metadata }) => {
   return (
