+++ components/index.js
@@ -6,8 +6,7 @@ export { ContactCardHome } from "./cards/ContactCardHome";
 
 export { CompanyLogos } from "./CompanyLogos";
 export { Logo } from "./Logo";
-export { PostMetadata } from "./post/PostMetadata";
-export { PostListItem } from "./post/PostListItem";
+export { CustomImage } from "./CustomImage";
 export { Summary } from "./Summary";
 export { Dropdown, DropdownItem } from "./Dropdown";
 export { Book } from "./Book";
@@ -15,10 +14,12 @@ export { Header } from "./Header";
 export { Hero } from "./Hero";
 export { Footer } from "./Footer";
 export { ButtonCTA } from "./ButtonCTA";
-export { NavigationLinks } from "./NavigationLinks";
+export { SectionLinks } from "./SectionLinks";
 export { Tidycal } from "./Tidycal";
 export { OpenPR } from "./OpenPR";
 export { Disqus } from "./Disqus";
+export { PostMetadata } from "./PostMetadata";
+export { PostListItem } from "./PostListItem";
 
 export * as elements from "./elements";
 export { MDXComponents } from "./MDXComponents";
